+++ src/main/java/school/faang/user_service/controller/user/UserController.java
@@ -3,22 +3,19 @@ package school.faang.user_service.controller.user;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.validator.constraints.Length;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.hibernate.validator.constraints.Length;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -31,12 +28,12 @@ import org.springframework.web.multipart.MultipartFile;
 import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.dto.user.UserIdsDto;
-import school.faang.user_service.dto.userJira.UserJiraCreateUpdateDto;
 import school.faang.user_service.dto.user.UserProfilePicDto;
+import school.faang.user_service.dto.user_jira.UserJiraCreateUpdateDto;
+import school.faang.user_service.dto.user_jira.UserJiraDto;
 import school.faang.user_service.entity.UserProfilePic;
 import school.faang.user_service.mapper.user.UserProfilePicMapper;
 import school.faang.user_service.service.avatar.AvatarService;
-import school.faang.user_service.dto.userJira.UserJiraDto;
 import school.faang.user_service.service.user.UserService;
 
 import java.util.List;
