+++ src/test/java/school/faang/user_service/service/user/UserServiceTest.java
@@ -9,21 +9,21 @@ import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.dto.user.UserFilterDto;
-import school.faang.user_service.dto.userJira.UserJiraCreateUpdateDto;
-import school.faang.user_service.dto.userJira.UserJiraDto;
+import school.faang.user_service.dto.user_jira.UserJiraCreateUpdateDto;
+import school.faang.user_service.dto.user_jira.UserJiraDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.userJira.UserJira;
 import school.faang.user_service.exception.EntityNotFoundException;
 import school.faang.user_service.exception.ErrorMessage;
 import school.faang.user_service.mapper.user.UserMapper;
-import school.faang.user_service.mapper.userJira.UserJiraMapper;
+import school.faang.user_service.mapper.user_jira.UserJiraMapper;
 import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.entity.premium.Premium;
 import school.faang.user_service.filter.user.UserEmailFilter;
 import school.faang.user_service.filter.user.UserFilter;
 import school.faang.user_service.filter.user.UserNameFilter;
 import school.faang.user_service.mapper.user.UserMapperImpl;
-import school.faang.user_service.service.userJira.UserJiraService;
+import school.faang.user_service.service.user_jira.UserJiraService;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
